REFACTOR: remove slider container and fetch data on the page
