+++ components/Slider.tsx
